checkPointForBackend

+++ src/Controller/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
     const videos = await Video.find({}).sort({ createdAt: "desc" });
@@ -8,7 +9,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id)
+    const video = await Video.findById(id).populate("owner")
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
@@ -19,17 +20,26 @@ export const watch = async (req, res) => {
 export const getEdit = async (req, res) => {
     const { id } = req.params;
     const video = await Video.findById(id);
+    const { user: { _id } } = req.session
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+
+    if (String(video.owner) !== String(_id)) {    //video.owner.toString()
+        return res.status(403).redirect("/")
     }
     return res.render("edit", { pageTitle: `Editing`, video });
 };
 export const postEdit = async (req, res) => {
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
-    const video = await Video.exists({ _id: id });
+    const video = await Video.findById(id);
+    const { user: { _id } } = req.session
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {    //video.owner.toString()
+        return res.status(403).redirect("/")
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -42,15 +52,21 @@ export const getUpload = (req, res) => {
     return res.render("upload", { pageTitle: "Upload Video" });
 };
 export const postUpload = async (req, res) => {
+    const { user: { _id } } = req.session
     const { title, description, hashtags } = req.body;
+    const { path: fileUrl } = req.file;
 
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
-
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id)
+        user.save();
         return res.redirect("/");
     } catch (error) {
         return res.status(400).render("upload", { pageTitle: "Upload Video", errorMessage: error._message });
@@ -59,7 +75,19 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session
+    const video = await Video.findById(id);
+    const user = await User.findById(_id);
+
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {    //video.owner.toString()
+        return res.status(403).redirect("/")
+    }
     await Video.findByIdAndDelete(id)
+    user.videos.splice(user.videos.indexOf(id), 1);
+    user.save();
     return res.redirect("/")
 }
